Further work on CLI guides
