+++ docs/guides/cli/authentication.md
@@ -1,0 +1,1 @@
+# Authentication
