+++ docs/guides/cli/hooks.md
@@ -1,0 +1,5 @@
+---
+outline: deep
+---
+
+# Hooks
