+++ docs/guides/cli/logging.md
@@ -1,0 +1,1 @@
+# Logging
