+++ docs/guides/cli/schemas.md
