+++ packages/cli/src/app/templates/config.tpl.ts
@@ -1,4 +1,5 @@
 import { generator, toFile, writeJSON } from '@feathershq/pinion'
+import { renderSource } from '../../commons'
 import { AppGeneratorContext } from '../index'
 
 const defaultConfig = ({}: AppGeneratorContext) => ({
@@ -24,8 +25,60 @@ const testConfig = {
   port: 8998
 }
 
+const configurationJsonTemplate =
+  ({}: AppGeneratorContext) => /* ts */ `// For more information about this file see https://dove.feathersjs.com/guides/cli/config.html#configuration-schemas
+import { defaultAppSettings, getValidator } from '@feathersjs/schema'
+import type { FromSchema } from '@feathersjs/schema'
+
+import { dataValidator } from './validators'
+
+export const configurationSchema = {
+  type: 'object',
+  additionalProperties: false,
+  required: [ 'host', 'port', 'public' ],
+  properties: {
+    ...defaultAppSettings,
+    host: { type: 'string' },
+    port: { type: 'number' },
+    public: { type: 'string' }
+  }
+} as const
+
+export const configurationValidator = getValidator(configurationSchema, dataValidator)
+
+export type ApplicationConfiguration = FromSchema<typeof configurationSchema>
+`
+
+const configurationTypeboxTemplate =
+  ({}: AppGeneratorContext) => /* ts */ `// For more information about this file see https://dove.feathersjs.com/guides/cli/config.html#configuration-schemas
+  import { Type, getValidator, defaultAppConfiguration } from '@feathersjs/typebox'
+import type { Static } from '@feathersjs/typebox'
+
+import { dataValidator } from './validators'
+
+export const configurationSchema = Type.Intersect([
+  defaultAppConfiguration,
+  Type.Object({
+    host: Type.String(),
+    port: Type.Number(),
+    public: Type.String()
+  })
+])
+
+export type ApplicationConfiguration = Static<typeof configurationSchema>
+
+export const configurationValidator = getValidator(configurationSchema, dataValidator)
+`
+
 export const generate = (ctx: AppGeneratorContext) =>
   generator(ctx)
     .then(writeJSON(defaultConfig, toFile('config', 'default.json')))
     .then(writeJSON(testConfig, toFile('config', 'test.json')))
     .then(writeJSON(customEnvironment, toFile('config', 'custom-environment-variables.json')))
+    .then(
+      renderSource(
+        async (ctx) =>
+          ctx.schema === 'typebox' ? configurationTypeboxTemplate(ctx) : configurationJsonTemplate(ctx),
+        toFile<AppGeneratorContext>(({ lib }) => lib, 'schemas', 'configuration')
+      )
+    )
