+++ packages/cli/src/app/templates/declarations.tpl.ts
@@ -3,7 +3,8 @@ import { AppGeneratorContext } from '../index'
 
 const template = ({
   framework
-}: AppGeneratorContext) => /* ts */ `import { HookContext as FeathersHookContext, NextFunction } from '@feathersjs/feathers'
+}: AppGeneratorContext) => /* ts */ `// For more information about this file see https://dove.feathersjs.com/guides/cli/typescript.html
+import { HookContext as FeathersHookContext, NextFunction } from '@feathersjs/feathers'
 import { Application as FeathersApplication } from '@feathersjs/${framework}'
 import { ApplicationConfiguration } from './schemas/configuration'
