+++ packages/cli/src/app/templates/services.tpl.ts
@@ -2,7 +2,9 @@ import { generator, toFile } from '@feathershq/pinion'
 import { renderSource } from '../../commons'
 import { AppGeneratorContext } from '../index'
 
-const template = ({}: AppGeneratorContext) => /* ts */ `import type { Application } from '../declarations'
+const template =
+  ({}: AppGeneratorContext) => /* ts */ `// For more information about this file see https://dove.feathersjs.com/guides/cli/application.html#configure-functions
+import type { Application } from '../declarations'
 
 export const services = (app: Application) => {
   // All services will be registered here
